move Address and TradeContact model and test to interpreter package
